Extract field type from filename
